+++ ngap/handler.go
@@ -1522,7 +1522,7 @@ func HandleInitialUEMessage(ran *context.AmfRan, message *ngapType.NGAPPDU, sctp
 			} else {
 				ranUe.Log.Tracef("find AmfUe [GUTI: %s]", guti)
 				/* checking the guti-ue belongs to this amf instance */
-				id, err := amfSelf.Drsm.FindOwnerInt32ID(int32(amfUe.Tmsi))
+				id, err := amfSelf.Drsm.FindOwnerInt32ID(amfUe.Tmsi)
 				if err != nil {
 					ranUe.Log.Errorf("Error checking the guti-ue in this instance: %v", err)
 				}
